requirements.txt
lightgbm
python-Levenshtein
pandas
numpy
scikit-learn
python-dateutil
xgboost
spacy
scipy
fuzzywuzzy
pytesseract
pdf2image
pillow
chardet
PyMuPDF
python-docx
docx2txt
streamlit
pdfplumber
opencv-python
setuptools
PyMuPDF
paddleocr
joblib
fpdf2
matplotlib
reportlab
faker
dill
px
gspread
oauth2client
plotly
cryptography

# paddlepaddle special install command (run separately):
# pip install paddlepaddle -i https://mirror.baidu.com/pypi/simple

transformers
torch
seaborn
weasyprint
kaleido
psycopg2-binary
supabase
werkzeug
passlib[bcrypt]
PyGObject
python-poppler
pycario